movie review showing

+++ database/rating.sql
@@ -4,7 +4,7 @@ CREATE TABLE IF NOT EXISTS movie_reviews (
     id INT AUTO_INCREMENT PRIMARY KEY,
     movie_id INT NOT NULL,
     user_id INT NOT NULL,
-    rating FLOAT NOT NULL,
+    rating INT NOT NULL,
     review TEXT,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     UNIQUE KEY unique_user_movie (movie_id, user_id), -- Unique constraint for user and movie combination
